Fixeamos size_t en i y j por int, debido loop inf

+++ main.c
@@ -6,10 +6,10 @@
 #define BITMAP_SIZE 16
 
 int first_fit(unsigned char *bitmap, size_t bitmap_size, size_t units_needed) {
-    size_t consecutive_free_bits = 0;
+    int consecutive_free_bits = 0;
     int first_free_bit_index = -1;
 
-    for (size_t i = 0; i < bitmap_size * 8; i++) {
+    for (int i = 0; i < bitmap_size * 8; i++) {
         int byte = i / 8;
         int bit = i % 8;
 
@@ -19,7 +19,7 @@ int first_fit(unsigned char *bitmap, size_t bitmap_size, size_t units_needed) {
             }
             consecutive_free_bits++;
             if (consecutive_free_bits == units_needed) {
-                for (size_t j = 0; j < units_needed; j++) {
+                for (int j = 0; j < units_needed; j++) {
                     int byte_to_set = (first_free_bit_index + j) / 8;
                     int bit_to_set = (first_free_bit_index + j) % 8;
                     bitmap[byte_to_set] |= (1 << (7 - bit_to_set));
@@ -35,8 +35,8 @@ int first_fit(unsigned char *bitmap, size_t bitmap_size, size_t units_needed) {
 }
 
 void print_bitmap(unsigned char *bitmap, size_t bitmap_size) {
-    for (size_t i = 0; i < bitmap_size; i++) {
-        for (size_t j = 7; j >= 0; j--) {
+    for (int i = 0; i < bitmap_size; i++) {
+        for (int j = 7; j >= 0; j--) {
             printf("%d", (bitmap[i] >> j) & 1);
         }
         if (i < bitmap_size - 1) {
